requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
openai
faiss-cpu
htmltemplate
sentence-transformers
langchain-community
InstructorEmbedding
streamlit-chat
transformers